+++ src/logic/wallets/ethTransactions.ts
@@ -1,12 +1,18 @@
 import { EthAdapterTransaction } from '@gnosis.pm/safe-core-sdk'
+import { GasPriceOracle } from '@gnosis.pm/safe-react-gateway-sdk'
 import axios from 'axios'
 import { BigNumber } from 'bignumber.js'
+import { FeeHistoryResult } from 'web3-eth'
+import { hexToNumber } from 'web3-utils'
+
 import { getSDKWeb3ReadOnly, getWeb3, getWeb3ReadOnly } from 'src/logic/wallets/getWeb3'
 import { getFixedGasPrice, getGasPriceOracles } from 'src/config'
-import { CodedException, Errors } from '../exceptions/CodedException'
-import { GasPriceOracle } from '@gnosis.pm/safe-react-gateway-sdk'
+import { CodedException, Errors, logError } from 'src/logic/exceptions/CodedException'
 
 export const EMPTY_DATA = '0x'
+
+const DEFAULT_MAX_GAS_FEE = 3.5e9 // 3.5 GWEI
+const DEFAULT_MAX_PRIO_FEE = 2.5e9 // 2.5 GWEI
 
 const fetchGasPrice = async (gasPriceOracle: GasPriceOracle): Promise<string> => {
   const { uri, gasParameter, gweiFactor } = gasPriceOracle
@@ -18,6 +24,44 @@ const fetchGasPrice = async (gasPriceOracle: GasPriceOracle): Promise<string> =>
     throw new Error('Fetched gas price is NaN')
   }
   return gasPrice.toString()
+}
+
+export const setMaxPrioFeePerGas = (maxPriorityFeePerGas: number, maxFeePerGas: number): number => {
+  return maxPriorityFeePerGas > maxFeePerGas ? maxFeePerGas : maxPriorityFeePerGas
+}
+
+export const getFeesPerGas = async (): Promise<{
+  maxFeePerGas: number
+  maxPriorityFeePerGas: number
+}> => {
+  let blocks: FeeHistoryResult | undefined
+  let maxPriorityFeePerGas: number | undefined
+  let baseFeePerGas: number | undefined
+
+  const web3 = getWeb3ReadOnly()
+
+  try {
+    // Lastest block, 50th reward percentile
+    blocks = await web3.eth.getFeeHistory(1, 'latest', [50])
+
+    // hexToNumber can throw if not parsing a valid hex string
+    baseFeePerGas = hexToNumber(blocks.baseFeePerGas[0])
+    maxPriorityFeePerGas = hexToNumber(blocks.reward[0][0])
+  } catch (err) {
+    logError(Errors._618, err.message)
+  }
+
+  if (!blocks || !maxPriorityFeePerGas || isNaN(maxPriorityFeePerGas) || !baseFeePerGas || isNaN(baseFeePerGas)) {
+    return {
+      maxFeePerGas: DEFAULT_MAX_GAS_FEE,
+      maxPriorityFeePerGas: DEFAULT_MAX_PRIO_FEE,
+    }
+  }
+
+  return {
+    maxFeePerGas: baseFeePerGas + maxPriorityFeePerGas,
+    maxPriorityFeePerGas,
+  }
 }
 
 export const calculateGasPrice = async (): Promise<string> => {
